pyproject.toml
[tool.poetry]
name = "mountain3d"
version = "0.1.0"
description = "3D визуализация зон обзора РЛС и зон огня ЗРК"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
plotly = "^5.16.1"
numpy = "^1.25.2"
dash = "^2.12.1"
dash-bootstrap-components = "^1.4.2"
geopy = "^2.3.0"
rasterio = "^1.3.8"
more-itertools = "^10.1.0"
numpy-stl = "^3.0.1"
pandas = "^2.0.3"
cx-freeze = "^7.2.0"
pydantic = "^2.8.2"
tqdm = "^4.66.5"
requests = "^2.32.3"
loguru = "^0.7.2"

[tool.poetry.dev-dependencies]
pre-commit = "^3.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
